+++ packages/react-querybuilder/src/utils/formatQuery/formatQuery.test.ts
@@ -1,4 +1,4 @@
-import type { RuleGroupType, RuleType } from '../../types/index.noReact';
+import type { FormatQueryOptions, RuleGroupType, RuleType } from '../../types/index.noReact';
 import { convertToIC } from '../convertQuery';
 import { prepareRuleGroup } from '../prepareQueryObjects';
 import { formatQuery } from './formatQuery';
@@ -54,8 +54,7 @@ it('handles json_without_ids correctly', () => {
 
 describe('parseNumbers', () => {
   it('parses numbers for json', () => {
-    expect(formatQuery(queryForNumberParsing, { format: 'json', parseNumbers: true })).toBe(
-      `{
+    const allNumbersParsed = `{
   "combinator": "and",
   "rules": [
     {
@@ -163,36 +162,82 @@ describe('parseNumbers', () => {
       ]
     }
   ]
-}`
-    );
+}`;
+    for (const opts of [
+      { parseNumbers: true },
+      { parseNumbers: 'strict' },
+      { parseNumbers: 'strict-limited', fields: [{ name: 'f', label: 'f', inputType: 'number' }] },
+    ] as FormatQueryOptions[]) {
+      expect(formatQuery(queryForNumberParsing, { ...opts, format: 'json' })).toBe(
+        allNumbersParsed
+      );
+    }
   });
 
   it('parses numbers for json_without_ids', () => {
-    expect(
-      JSON.parse(
-        formatQuery(prepareRuleGroup(queryForNumberParsing), {
-          format: 'json_without_ids',
-          parseNumbers: true,
-        })
-      )
-    ).toEqual(
-      JSON.parse(
-        '{"rules":[{"field":"f","value":"NaN","operator":">"},{"field":"f","value":0,"operator":"="},{"field":"f","value":0,"operator":"="},{"field":"f","value":0,"operator":"="},{"rules":[{"field":"f","value":1.5,"operator":"<"},{"field":"f","value":1.5,"operator":">"}],"combinator":"or"},{"field":"f","value":[0,1,2],"operator":"in"},{"field":"f","value":[0,1,2],"operator":"in"},{"field":"f","value":"0, abc, 2","operator":"in"},{"field":"f","value":[0,1],"operator":"between"},{"field":"f","value":[0,1],"operator":"between"},{"field":"f","value":"0, abc","operator":"between"},{"field":"f","value":1,"operator":"between"},{"field":"f","value":1,"operator":"between"},{"field":"f","value":[1],"operator":"between"},{"field":"f","value":[{},{}],"operator":"between"}],"combinator":"and"}'
-      )
+    const allNumbersParsed = JSON.parse(
+      '{"rules":[{"field":"f","value":"NaN","operator":">"},{"field":"f","value":0,"operator":"="},{"field":"f","value":0,"operator":"="},{"field":"f","value":0,"operator":"="},{"rules":[{"field":"f","value":1.5,"operator":"<"},{"field":"f","value":1.5,"operator":">"}],"combinator":"or"},{"field":"f","value":[0,1,2],"operator":"in"},{"field":"f","value":[0,1,2],"operator":"in"},{"field":"f","value":"0, abc, 2","operator":"in"},{"field":"f","value":[0,1],"operator":"between"},{"field":"f","value":[0,1],"operator":"between"},{"field":"f","value":"0, abc","operator":"between"},{"field":"f","value":1,"operator":"between"},{"field":"f","value":1,"operator":"between"},{"field":"f","value":[1],"operator":"between"},{"field":"f","value":[{},{}],"operator":"between"}],"combinator":"and"}'
     );
+    for (const opts of [
+      { parseNumbers: true },
+      { parseNumbers: 'strict' },
+      { parseNumbers: 'strict-limited', fields: [{ name: 'f', label: 'f', inputType: 'number' }] },
+    ] as FormatQueryOptions[]) {
+      expect(
+        JSON.parse(
+          formatQuery(prepareRuleGroup(queryForNumberParsing), {
+            ...opts,
+            format: 'json_without_ids',
+          })
+        )
+      ).toEqual(allNumbersParsed);
+    }
   });
 
   it('parses numbers for json_without_ids with independentCombinators', () => {
+    const allNumbersParsed = JSON.parse(
+      '{"rules":[{"field":"f","value":"NaN","operator":">"},"and",{"field":"f","value":0,"operator":"="},"and",{"field":"f","value":0,"operator":"="},"and",{"field":"f","value":0,"operator":"="},"and",{"rules":[{"field":"f","value":1.5,"operator":"<"},"or",{"field":"f","value":1.5,"operator":">"}]},"and",{"field":"f","value":[0,1,2],"operator":"in"},"and",{"field":"f","value":[0,1,2],"operator":"in"},"and",{"field":"f","value":"0, abc, 2","operator":"in"},"and",{"field":"f","value":[0,1],"operator":"between"},"and",{"field":"f","value":[0,1],"operator":"between"},"and",{"field":"f","value":"0, abc","operator":"between"},"and",{"field":"f","value":1,"operator":"between"},"and",{"field":"f","value":1,"operator":"between"},"and",{"field":"f","value":[1],"operator":"between"},"and",{"field":"f","value":[{},{}],"operator":"between"}]}'
+    );
+    for (const opts of [
+      { parseNumbers: true },
+      { parseNumbers: 'strict' },
+      { parseNumbers: 'strict-limited', fields: [{ name: 'f', label: 'f', inputType: 'number' }] },
+    ] as FormatQueryOptions[]) {
+      expect(
+        JSON.parse(
+          formatQuery(prepareRuleGroup(convertToIC(queryForNumberParsing)), {
+            ...opts,
+            format: 'json_without_ids',
+          })
+        )
+      ).toEqual(allNumbersParsed);
+    }
+  });
+
+  it('parses numbers only when inputType is number', () => {
     expect(
       JSON.parse(
-        formatQuery(prepareRuleGroup(convertToIC(queryForNumberParsing)), {
-          format: 'json_without_ids',
-          parseNumbers: true,
-        })
+        formatQuery(
+          {
+            rules: [
+              { field: 'f1', operator: '=', value: '123' },
+              'and',
+              { field: 'f2', operator: '=', value: '123' },
+            ],
+          },
+          {
+            format: 'json_without_ids',
+            parseNumbers: 'strict-limited',
+            fields: [
+              { name: 'f1', label: 'f1', inputType: 'number' },
+              { name: 'f2', label: 'f2' },
+            ],
+          }
+        )
       )
     ).toEqual(
       JSON.parse(
-        '{"rules":[{"field":"f","value":"NaN","operator":">"},"and",{"field":"f","value":0,"operator":"="},"and",{"field":"f","value":0,"operator":"="},"and",{"field":"f","value":0,"operator":"="},"and",{"rules":[{"field":"f","value":1.5,"operator":"<"},"or",{"field":"f","value":1.5,"operator":">"}]},"and",{"field":"f","value":[0,1,2],"operator":"in"},"and",{"field":"f","value":[0,1,2],"operator":"in"},"and",{"field":"f","value":"0, abc, 2","operator":"in"},"and",{"field":"f","value":[0,1],"operator":"between"},"and",{"field":"f","value":[0,1],"operator":"between"},"and",{"field":"f","value":"0, abc","operator":"between"},"and",{"field":"f","value":1,"operator":"between"},"and",{"field":"f","value":1,"operator":"between"},"and",{"field":"f","value":[1],"operator":"between"},"and",{"field":"f","value":[{},{}],"operator":"between"}]}'
+        '{"rules":[{"field":"f1","operator":"=","value":123},"and",{"field":"f2","operator":"=","value":"123"}]}'
       )
     );
   });